requirements.txt
numpy == 1.19.5
pandas == 1.2.1
psycopg2_binary == 2.8.6
